ARM: meson: Add boot over DFU RAM as an USB boot step
Add boot over DFU RAM as an alternate to running script at
a fixed address like done today.

The main culprit is that it's not possible to do that
on G12A/SM1 platforms due to changes in the USB boot protocol.

With this, U-Boot will present a DFU device with a ram slot where
the Host could write a fitImage or legacy U-Boot image, then with the
detach command boot will continue trying to boot the uploaded image.

To: Christian Hewitt <christianshewitt@gmail.com>
To: Vyacheslav Bocharov <adeep@lexina.in>
To: Beniamino Galvani <b.galvani@gmail.com>
Cc: u-boot-amlogic@groups.io
Cc: u-boot@lists.denx.de
Signed-off-by: Neil Armstrong <neil.armstrong@linaro.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20231023-usb-dfu-boot-v1-0-df9d121c67c1@linaro.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20231023-usb-dfu-boot-f47849f9fb34",
    "prefixes": [],
    "history": {
      "v1": [
        "20231023-usb-dfu-boot-v1-0-df9d121c67c1@linaro.org"
      ]
    }
  }
}